Conflicts: (trivially solved)
	ADOL-C/examples/additional_examples/sparse/sparse_hessian.cpp
	ADOL-C/examples/additional_examples/sparse/sparse_jacobian.cpp

+++ ADOL-C/examples/additional_examples/lighthouse/cubic-iter-2.cpp
@@ -66,7 +66,7 @@ int main() {
     trace_off();
 
     /*--------------------------------------------------------------------------*/
-    int tape_stats[STAT_SIZE];
+    size_t tape_stats[STAT_SIZE];
     tapestats(tag,tape_stats);
 
     fprintf(stdout,"\n    independents            %d\n",tape_stats[NUM_INDEPENDENTS]);
